// src/models/paciente.ts
export class Paciente {
    nome : string;
    nomeMae : string;
    dataNascimento : Date;
    sexo: number;
    raca: number;
    cep: string;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    uf: string;
    telefone : string;
    email: string;
}